+++ tests/testthat/test-check_metadata_matches_schema.R
@@ -29,3 +29,54 @@ test_that("check_metadata_matches_schema works", {
     )
   )
 })
+
+test_that("check_metadata_matches_schema properly distinguishes length-1 arrays from scalars", {
+  hub_path <- test_path("testdata", "array_hub")
+
+  run_check <- function(file_path, expected_result, expected_message_regex) {
+    result <- check_metadata_matches_schema(
+      file_path = file_path,
+      hub_path = hub_path
+    )
+
+    expect_s3_class(result, expected_result)
+    expect_match(result$message, expected_message_regex)
+  }
+
+  # multi-entry array passes
+  run_check("multi-entry.yml", "check_success", "consistent with")
+
+  # length-1 array passes
+  run_check("single-entry.yml", "check_success", "consistent with")
+
+  # correct length one optional array passes
+  run_check("two-arrays.yml", "check_success", "consistent with")
+
+  # missing mandatory array field errors
+  run_check(
+    "no-entry.yml",
+    "check_error",
+    "must have required property"
+  )
+
+  # null mandatory array field errors
+  run_check(
+    "null-entry.yml",
+    "check_error",
+    "mandatory_array_valued_metadata must be array"
+  )
+
+  # string in an array field errors, even if optional field
+  run_check(
+    "string-instead-of-array.yml",
+    "check_error",
+    "optional_array_valued_metadata must be array"
+  )
+
+  # length-1 array in string field errors, even if optional field
+  run_check(
+    "array-instead-of-string.yml",
+    "check_error",
+    "website_url must be string"
+  )
+})